+++ packages/flet/lib/src/controls/create_control.dart
@@ -335,6 +335,7 @@ Widget createWidget(
           control: controlView.control,
           children: controlView.children,
           parentDisabled: parentDisabled,
+          parentAdaptive: parentAdaptive,
           backend: backend);
     case "textbutton":
       return TextButtonControl(