+++ packages/flet/lib/src/controls/cupertino_textfield.dart
@@ -50,9 +50,10 @@ class _CupertinoTextFieldControlState extends State<CupertinoTextFieldControl> {
     super.initState();
     _controller = TextEditingController();
     _shiftEnterfocusNode = FocusNode(
-      onKey: (FocusNode node, RawKeyEvent evt) {
-        if (!evt.isShiftPressed && evt.logicalKey.keyLabel == 'Enter') {
-          if (evt is RawKeyDownEvent) {
+      onKeyEvent: (FocusNode node, KeyEvent evt) {
+        if (!HardwareKeyboard.instance.isShiftPressed &&
+            evt.logicalKey.keyLabel == 'Enter') {
+          if (evt is KeyDownEvent) {
             widget.backend.triggerControlEvent(widget.control.id, "submit");
           }
           return KeyEventResult.handled;
@@ -80,7 +81,8 @@ class _CupertinoTextFieldControlState extends State<CupertinoTextFieldControl> {
     setState(() {
       _focused = _shiftEnterfocusNode.hasFocus;
     });
-    widget.backend.triggerControlEvent(widget.control.id, _shiftEnterfocusNode.hasFocus ? "focus" : "blur");
+    widget.backend.triggerControlEvent(widget.control.id,
+        _shiftEnterfocusNode.hasFocus ? "focus" : "blur", "");
   }
 
   void _onFocusChange() {
@@ -88,7 +90,7 @@ class _CupertinoTextFieldControlState extends State<CupertinoTextFieldControl> {
       _focused = _focusNode.hasFocus;
     });
     widget.backend.triggerControlEvent(
-        widget.control.id, _focusNode.hasFocus ? "focus" : "blur");
+        widget.control.id, _focusNode.hasFocus ? "focus" : "blur", "");
   }
 
   @override
@@ -212,7 +214,8 @@ class _CupertinoTextFieldControlState extends State<CupertinoTextFieldControl> {
         enabled: !disabled,
         onSubmitted: !multiline
             ? (_) {
-                widget.backend.triggerControlEvent(widget.control.id, "submit");
+                widget.backend
+                    .triggerControlEvent(widget.control.id, "submit", "");
               }
             : null,
         decoration: defaultDecoration?.copyWith(