+++ packages/flet/lib/src/controls/elevated_button.dart
@@ -8,6 +8,7 @@ import '../utils/icons.dart';
 import '../utils/launch_url.dart';
 import 'create_control.dart';
 import 'cupertino_button.dart';
+import 'cupertino_dialog_action.dart';
 import 'error.dart';
 import 'flet_store_mixin.dart';
 
@@ -66,12 +67,21 @@ class _ElevatedButtonControlState extends State<ElevatedButtonControl>
       if (adaptive == true &&
           (platform == TargetPlatform.iOS ||
               platform == TargetPlatform.macOS)) {
-        return CupertinoButtonControl(
-            control: widget.control,
-            parentDisabled: widget.parentDisabled,
-            parentAdaptive: adaptive,
-            children: widget.children,
-            backend: widget.backend);
+        return widget.control.name == "action" &&
+                (widget.parent?.type == "alertdialog" ||
+                    widget.parent?.type == "cupertinoalertdialog")
+            ? CupertinoDialogActionControl(
+                control: widget.control,
+                parentDisabled: widget.parentDisabled,
+                parentAdaptive: adaptive,
+                children: widget.children,
+                backend: widget.backend)
+            : CupertinoButtonControl(
+                control: widget.control,
+                parentDisabled: widget.parentDisabled,
+                parentAdaptive: adaptive,
+                children: widget.children,
+                backend: widget.backend);
       }
 
       String text = widget.control.attrString("text", "")!;