+++ packages/flet/lib/src/controls/pagelet.dart
@@ -67,28 +67,15 @@ class _PageletControlState extends State<PageletControl> with FletStoreMixin {
     }
 
     return withPagePlatform((context, platform) {
-      PageDesign widgetsDesign = PageDesign.material;
-
-      var pageDesign = PageDesign.values.firstWhereOrNull((a) =>
-          a.name.toLowerCase() ==
-          widget.control.attrString("design", "")!.toLowerCase());
-
-      if ((pageDesign == PageDesign.adaptive &&
-              (platform == TargetPlatform.iOS ||
-                  platform == TargetPlatform.macOS)) ||
-          pageDesign == PageDesign.cupertino) {
-        widgetsDesign = PageDesign.cupertino;
-      } else {
-        widgetsDesign = PageDesign.material;
-      }
-
       bool disabled = widget.control.isDisabled || widget.parentDisabled;
-      bool? adaptive = pageDesign != null
-          ? pageDesign == PageDesign.adaptive
-          : widget.parentAdaptive;
+      bool? adaptive =
+          widget.control.attrBool("adaptive") ?? widget.parentAdaptive;
 
-      var bgcolor = HexColor.fromString(
-          Theme.of(context), widget.control.attrString("bgcolor", "")!);
+      var widgetsDesign = adaptive == true &&
+              (platform == TargetPlatform.iOS ||
+                  platform == TargetPlatform.macOS)
+          ? PageDesign.cupertino
+          : PageDesign.material;
 
       List<String> childIds = [
         appBarCtrls.firstOrNull?.id,
@@ -217,13 +204,11 @@ class _PageletControlState extends State<PageletControl> with FletStoreMixin {
                 : null;
 
             Widget scaffold = Scaffold(
-                key: widgetsDesign != PageDesign.cupertino ? scaffoldKey : null,
+                key: bar == null || bar is AppBarControl ? scaffoldKey : null,
                 backgroundColor: HexColor.fromString(Theme.of(context),
                         widget.control.attrString("bgcolor", "")!) ??
                     CupertinoTheme.of(context).scaffoldBackgroundColor,
-                appBar: widgetsDesign != PageDesign.cupertino
-                    ? bar as PreferredSizeWidget?
-                    : null,
+                appBar: bar is AppBarControl ? bar : null,
                 drawer: drawerView != null
                     ? NavigationDrawerControl(
                         control: drawerView.control,
@@ -269,7 +254,7 @@ class _PageletControlState extends State<PageletControl> with FletStoreMixin {
                     : null,
                 floatingActionButtonLocation: fabLocation);
 
-            if (widgetsDesign == PageDesign.cupertino) {
+            if (bar is CupertinoAppBarControl) {
               scaffold = CupertinoPageScaffold(
                   key: scaffoldKey,
                   backgroundColor: HexColor.fromString(Theme.of(context),