+++ packages/flet/lib/src/controls/textfield.dart
@@ -48,9 +48,10 @@ class _TextFieldControlState extends State<TextFieldControl>
     super.initState();
     _controller = TextEditingController();
     _shiftEnterfocusNode = FocusNode(
-      onKey: (FocusNode node, RawKeyEvent evt) {
-        if (!evt.isShiftPressed && evt.logicalKey.keyLabel == 'Enter') {
-          if (evt is RawKeyDownEvent) {
+      onKeyEvent: (FocusNode node, KeyEvent evt) {
+        if (!HardwareKeyboard.instance.isShiftPressed &&
+            evt.logicalKey.keyLabel == 'Enter') {
+          if (evt is KeyDownEvent) {
             widget.backend.triggerControlEvent(widget.control.id, "submit");
           }
           return KeyEventResult.handled;
@@ -78,7 +79,8 @@ class _TextFieldControlState extends State<TextFieldControl>
     setState(() {
       _focused = _shiftEnterfocusNode.hasFocus;
     });
-    widget.backend.triggerControlEvent(widget.control.id, _shiftEnterfocusNode.hasFocus ? "focus" : "blur");
+    widget.backend.triggerControlEvent(widget.control.id,
+        _shiftEnterfocusNode.hasFocus ? "focus" : "blur", "");
   }
 
   void _onFocusChange() {
@@ -86,7 +88,7 @@ class _TextFieldControlState extends State<TextFieldControl>
       _focused = _focusNode.hasFocus;
     });
     widget.backend.triggerControlEvent(
-        widget.control.id, _focusNode.hasFocus ? "focus" : "blur");
+        widget.control.id, _focusNode.hasFocus ? "focus" : "blur", "");
   }
 
   @override
@@ -230,7 +232,7 @@ class _TextFieldControlState extends State<TextFieldControl>
           onFieldSubmitted: !multiline
               ? (_) {
                   widget.backend
-                      .triggerControlEvent(widget.control.id, "submit");
+                      .triggerControlEvent(widget.control.id, "submit", "");
                 }
               : null,
           decoration: buildInputDecoration(