+++ packages/flet_video/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.20.2
+
+No changes in this release. Version bumped to follow parent `flet` package.
+
+## 0.20.1
+
+No changes in this release. Version bumped to follow parent `flet` package.
+
 ## 0.20.0
 
 Initial release of the package.